+++ App/Controllers/UsersController.php
@@ -6,6 +6,8 @@ use App\Core\AControllerBase;
 use App\Core\Responses\Response;
 use App\Models\User;
 use App\Models\Post;
+use App\Models\Order;
+use App\Models\Location;
 
 
 class UsersController extends AControllerBase
@@ -34,10 +36,23 @@ class UsersController extends AControllerBase
      */
     public function delete()
     {
-
-        $post = User::getOne($this->request()->getValue('id'));
-        $post->delete();
-
+        if ($_SESSION['user'] == "cekel1@stud.uniza.sk")
+        {
+            $user = User::getOne($this->request()->getValue('id'));
+            $posts = Post::getAll("id_user = ?", [$user->getId()]);
+            foreach($posts as $post)
+            {
+                $post->delete();
+            }
+            $orders = Order::getAll("id_user = ?", [$user->getId()]);
+            foreach($orders as $order)
+            {
+                $order->delete();
+            }
+            $loc = Location::getAll("id_user = ?", [$user->getId()]);
+            $loc[0]->delete();
+            $user->delete();
+        }
         return $this->redirect("?c=users");
     }
